1) Considering non targets estimate as real numbers in phase 1 instead of rounding them 2) Changing phase 2 accordingly to consider real numbers

+++ InfluenceMaximization/FunctionEvaluators.hpp
@@ -24,7 +24,7 @@ class TIMEvaluator {
 private:
     Graph *graph;
     EstimateNonTargets estimateNonTargets;
-    vector<int> nonTargets;
+    vector<double> nonTargets;
     vector<vector<int>> rrSetsTargets;
     vector<vector<int>> rrSetsNonTargets;
     TIMCoverage *timCoverage;
